+++ index.html
@@ -41,6 +41,6 @@
       
       <div class="movie-list" id="movie-list"></div>
     </main>
-    <script src="/app.js"></script>
+    <script type="module" src="js/main.js"></script>
   </body>
 </html>
