+++ watchlist.html
@@ -25,6 +25,6 @@
       <div class="watchlist" id="watchlist"></div>
     </main>
 
-    <script src="/app.js"></script>
+    <script type="module" src="js/main.js"></script>
   </body>
 </html>
